完善了UserController

+++ src/main/resources/ken.sql
@@ -14,7 +14,7 @@ create table ken_log
 -- 用户表
 CREATE TABLE user
 (
-	id BIGINT(20) NOT NULL COMMENT '主键ID',
+	id int NOT NULL COMMENT '主键ID',
 	name VARCHAR(30) NULL DEFAULT NULL COMMENT '姓名',
 	age INT(11) NULL DEFAULT NULL COMMENT '年龄',
 	email VARCHAR(50) NULL DEFAULT NULL COMMENT '邮箱',
